add usePorterYaml and app to proto conversion
